renamed plot module
